GWS DeleteGroup() and convenience methods. m
Implemented GoogleWorkspaceClient.DeleteGroup().
Added try catches to GoogleWorkspaceClient request execution calls.
Added GoogleWorkspaceClient.BuildServiceAccountCredential to make it easier to config the authentication for the client.
Added DirectoryApiScopes with a list of properties as defined by Google's Directory API to make it easier to define scopes the client should request from the API.