Added functionality to update Okta User profile attributes directly via JSON.
Galactic.Identity.Okta.OktaClient.UpdateUserProfile()
For working with custom attributes its sometimes easier to deal with the JSON directly as opposed to having to code up wrappers. This method allows for that.

This update also adds some additional HTTP status codes that Okta responds with, so that the client acts correctly.